move testutil's code to proper repo
