+++ service/block/event_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/celestiaorg/celestia-node/service/header"
-	"github.com/celestiaorg/celestia-node/testutils"
 	core "github.com/tendermint/tendermint/types"
 )
 
@@ -92,7 +91,7 @@ func (m *mockFetcher) generateBlocks(t *testing.T, num int) []Block {
 func generateRawAndExtendedBlock(t *testing.T) (*RawBlock, *Block) {
 	t.Helper()
 
-	data, err := testutils.GenerateRandomBlockData(1, 1, 1, 1, 40)
+	data, err := GenerateRandomBlockData(1, 1, 1, 1, 40)
 	if err != nil {
 		t.Fatal(err)
 	}
